Create Jenkinsfile
Jenkinsfile that is use to to run a series of command such as cloning the  project,  running tests on it and deploying it. For it to work on Jenkins, please have specify paths to at least  one JDK and Maven Installation, which you can specify in the Global Tool Configuration tab in Jenkins. Please name the maven path Maven 3.6.3 and the jdk path as java-1.8.0-openjdk. This pipeline will not work if the git repository was already cloned  in the respectively directory it will be in when this Jenkinsfile commands are executed. Essentially after every time you run this file, please delete the repo it clone at its respective location it was cloned.